Renamed motion.py and test_motion.py to kinematics.py and kinematicspy respectively.
