+++ movement/kinematics/__init__.py
@@ -1,15 +1,15 @@
-from .motion import (
+from .kinematics import (
     compute_displacement,
     compute_velocity,
     compute_acceleration,
     compute_time_derivative,
     compute_speed,
+    compute_path_length,
 )
 from .navigation import (
     compute_forward_vector,
     compute_head_direction_vector,
     compute_forward_vector_angle,
-    compute_path_length,
 )
 from .distances import compute_pairwise_distances
 
@@ -19,9 +19,9 @@ __all__ = [
     "compute_acceleration",
     "compute_time_derivative",
     "compute_speed",
+    "compute_path_length",
     "compute_forward_vector",
     "compute_head_direction_vector",
     "compute_forward_vector_angle",
-    "compute_path_length",
     "compute_pairwise_distances",
 ]
